refactor: move tab functions into repository layer
