--- src/ui/functions/groupTabsBySearchKeyword.ts
@@ -1,27 +1,0 @@
-import {
-  findGroupsByName,
-  flatTabsInWindows,
-  type Window,
-} from "../../model/Window";
-import { createGroupWithTabs } from "../../platform/repository/TabGroupRepository";
-import { addTabsToGroup } from "../../platform/repository/TabsRepository";
-
-const groupTabsBySearchKeyword = async (
-  keyword: string,
-  windows: Window[],
-  tabIds: number[],
-) => {
-  const pinnedTabs = flatTabsInWindows(windows).filter((tab) => tab.pinned);
-  const tabIdsExcludingPinned = tabIds.filter(
-    (tabId) => !pinnedTabs.some((tab) => tab.id === tabId),
-  );
-
-  const sameNameGroups = findGroupsByName(keyword, windows);
-  if (sameNameGroups.length > 0) {
-    await addTabsToGroup(tabIdsExcludingPinned, sameNameGroups[0].id);
-  }
-
-  await createGroupWithTabs(keyword, tabIdsExcludingPinned);
-};
-
-export default groupTabsBySearchKeyword;
