--- src/ui/functions/moveTabFromPinnedToPinned.ts
@@ -1,18 +1,0 @@
-import { moveTab, pinTab } from "../../platform/repository/TabsRepository";
-
-const moveTabFromPinnedToPinned = async (
-  tabId: number,
-  sourceWindowId: number,
-  destWindowId: number,
-  index: number,
-) => {
-  if (sourceWindowId === destWindowId) {
-    await moveTab(tabId, sourceWindowId, index);
-  } else {
-    await moveTab(tabId, destWindowId, -1);
-    await pinTab(tabId);
-    await moveTab(tabId, destWindowId, index);
-  }
-};
-
-export default moveTabFromPinnedToPinned;
