--- src/ui/functions/moveTabFromRootToPinned.ts
@@ -1,13 +1,0 @@
-import { moveTab, pinTab } from "../../platform/repository/TabsRepository";
-
-const moveTabFromRootToPinned = async (
-  tabId: number,
-  windowId: number,
-  index: number,
-) => {
-  await moveTab(tabId, windowId, -1);
-  await pinTab(tabId);
-  await moveTab(tabId, windowId, index);
-};
-
-export default moveTabFromRootToPinned;
