+++ src/ui/options/TabSearchForm.tsx
@@ -25,9 +25,9 @@ import {
   focusTab,
   getRecentActiveTabs,
 } from "../../platform/repository/TabsRepository";
+import { groupTabsBySearchKeyword } from "../../platform/repository/TabGroupRepository";
 import TabItem from "../components/TabItem";
 import { WindowsContext, WindowsProvider } from "../contexts/WindowsContext";
-import groupTabsBySearchKeyword from "../functions/groupTabsBySearchKeyword";
 
 const smoothScrollTo = (container: HTMLElement, targetScrollTop: number) => {
   const startScrollTop = container.scrollTop;
